aapt : Optimizations
aapt: Speed up the style pruning

Prior to this commit it removed each style one by one which
causes the vector to repeatedly be shrunk and reallocated and
copied (aka either quadratic or NlogN in the number of items
removed).

This commit simply makes it remove them all at once to make it
linear instead.

Change-Id: I541d151675ab19f37d9de1e7a323104d0d3b3c63

aapt: Use a std::map instead of a SortedVector

Android's SortedVectorImpl uses arrays that it must
insert into the middle of.  Each insertion is O(N) time
because it must move on average half the elements of
the array to make room for the new element.  That is,
O(N^2) time to build this sorted vector.

std::map on the other hand normally uses red/black
trees and has a cost of NlogN to add N elements to it.

Change-Id: I5da0363ba806ab615b2aad0fb2a43ef9a9bec327

aapt: Fix compiler warning (clobbered by longjmp)

The error handling can clobber the fp variable when the code is
being optimized and therefore fp must be marked as volatile to
stop it from doing unsafe optimizations.

Change-Id: Ib133587687c7fc0e788c4b9673669ef9b5bbe91a