hwui : Optimizations
Fix random crash in libhwui

On some cases (noticed from log) that there is a crash in
libhwui.so, future debug log shows that the light center position
values are nan, causing an out of bound access in hull function
in SpotShadow.cpp for parameter retPoly which assume to be of
maximum length of pointsLength only but not the case when all
points are nan.

More log shows that a Snapshot was created (new) without
mRelativeLightCenter begin set, and Layer.cpp trying to
use it by calling updateLightPosFromRenderer. Those values
are random values in stack which causing math calc based on
that value returning nan also.

Initialize mRelativeLightCenter to 0 in Snapshot.cpp to avoid the
problem.

Change-Id: I9e24f6f1f856d053d09db1b1aae100f219f2a6de

Fix memory corruption caused by patchCache.clear

The workfolw of drawPatch in HWUI is like this:
applyDraw -> getMesh -> patchCache.get -> GlopBuilder -> renderGlop
-> RenderState.render

If trimMemory is invoked before drawPatch, the content in mesh may
be corrupted by patchCache.clear. Meanwhile, mGenerationId is still
the same before and after clear operation, so mesh is not updated,
the wrong info is used in rendering, thus lead to driver crash.

Test: run com.immomo.momo app which reproduce the crash

CRs-fixed: 1088161
Change-Id: I62b4b0074068fd6e9789c3034c43ac7fcb4444d8

Fix a use-of-uninitialized-value warning.

This warning was generated by the static analyzer:
warning: The left expression of the compound assignment is an
uninitialized value. The computed value will also be garbage
        *mDirtyMask |= animator->dirtyMask();

This can be fixed by initializing dirtyMask before using it in
AnimateFunctor.

Bug: None
Test: Static analyzer no longer complains about this.
Change-Id: I892749079cf93a012f1bed059a3f0c8bbaed9d7e

Avoid starving RT anims

Test: Manual, usleep(16000) in DrawFrameTask and tap
on recents

Merged-In: I88bb30a2503bc908ec45650c7d36b6fb3cc750d0

Change-Id: I88bb30a2503bc908ec45650c7d36b6fb3cc750d0
(cherry picked from commit 622a96ad761e8736a6b4583015696a19387843e0)

Reduce hwui CPU time by using glDrawRangeElements

The CPU overhead of glDrawElements in the GPU driver is significant
with client-side vertex data (unique per draw call) as the driver has
to calculate the mininimum and maximum indices from the index buffer
in order to evaluate the range of vertex data required. This can be
avoided by keeping track of the min-max in hwui and passing it with
glDrawRangeElements. This requires OpenGL ES3.0 support (which is
already checked for elsewhere in hwui).

Test: manual - visual inspection on fugu (nexus player)

Change-Id: I57bb1ddd239a1032f74f1cd2683bbe0970e84bd9

Disable hwui blending for first draw to main FBO

In some applications, the first draw is not opaque - either because the
application is misbehaved, or because hwui is not able to reliably tell
whether the layer is opaque or translucent. This is undefined behaviour
in OpenGL ES and has a significant performance and bandwidth impact on
some tiler GPUs as it requires loading the previous frame's color data.
This change disables blending in that case and also for effectively
opaque blend modes (SRC=GL_ONE, DST=GL_ZERO). It increases performance
by ~10% for Leanback CTS on some low-end GPUs (gradient layer that hwui
incorrectly believes to be translucent).

Test: manual - visual inspection on fugu (nexus player)

Change-Id: I2cbf1c76678acae1a36923e72fd18ed55cd89dc2

Optimise identical hwui OpenGL ES programs

The hashing in Program.h did not fully match ProgramCache.cpp.
This caused OpenGL ES driver overhead when switching programs.
This change tracks the "single colour" fast path in Program.h
and also adds a special mode for "opaque" colours to allow
further driver optimisations.

Test: manual - visual inspection on fugu (nexus player)

Change-Id: Icc640dc195e44db76a12ea283d12ca4437c4d6e0

Fast paths for hwui matrix and gradient shaders

The shaders for matrix and simple gradients do many operations which
are not necessary for the most common values, and the matrix shader
is extremely shader heavy (causes noticeable slowdown on low-end GPUs
for e.g. the contacts application). This patch detects these common
inputs and uses much faster shaders instead. It also optimises the
un-premultiply and re-multiply for the general matrix shader, which
should in theory be as accurate (possibly more accurate).

Test: manual - visual inspection on fugu (nexus player)

Change-Id: I6b5fb4ee6f01aea37d09598d09ef86ee8135e39d

Optimised hwui shadow shader by using Vec2 alpha

The original fragment shader has unnecessary arithmetic instructions
on some GPUs due to using vec2(alpha, 0.5) instead of directly passing
the Vec2 texture coordinates. By passing the Vec2 directly, it improves
performance on low-end GPUs by letting them reach their peak fillrate.
It also optimises the multiplication as RGB are known to always be 0.

Test: manual - visual inspection on fugu (nexus player)

Change-Id: I438ff01f899153ac2e3d62ab13ac371fd2745df1

Optimised hwui rounded corners shader

This is a ~2x performance increase for the rounded corners shader which
was previously shader limited even on many high-end GPUs. It simplifies
the calculations by pre-dividing by the radius, which also allows to
maintain mediump/FP16 precision (this improves performance and might
improve image quality on GPUs that don't support highp e.g. Mali-400).
It also uses a conditional in the shader to save much of the work on
the vast majority of pixels which are not close to the corners.

Test: manual - visual inspection on fugu (nexus player)

Change-Id: I38f491cfb006e2991ef0b9d6b3d67d2c4d328d44

hwui: Remove deprecated compiler flag

 * HWUI_COMPILE_FOR_PERF uses the -mapcs flag
   but it has already been deprecated.
 * Remove the flag "-mapcs" completely

Reference:
https://gcc.gnu.org/onlinedocs/gcc/ARM-Options.html#ARM-Options

Change-Id: I6f72b0dc37deefbbfb1c99161f9577a9c447a343